feat(npu): Ascend NPU adaptation for DPA4/SeZM training
- deepmd/env.py: GLOBAL_ENER_FLOAT_PRECISION low->float32 (NPU matmul
  does not support double dtype)
- deepmd/pt/utils/env.py: add DEVICE=npu branch (transfer_to_npu +
  per-rank set_device for multi-card DDP)
- deepmd/pt/entrypoints/main.py: use hccl backend for DDP on NPU
- deepmd/pt/optimizer/hybrid_muon.py: skip torch.compile on NPU
  (transfer_to_npu + dynamo incompatible)
- deepmd/pt_expt/utils/vesin_graph_builder.py: force CPU neighbor
  search on NPU (vesin does not support NPU devices)
- deepmd/pt/model/descriptor/sezm_nn/utils.py: guard nvtx_range
  against RuntimeError on CPU-only torch + transfer_to_npu

Verified: DPA4 single-card and 2-card DDP training pass on Ascend910
with DP_INTERFACE_PREC=low.

+++ deepmd/env.py
@@ -40,7 +40,7 @@ if dp_float_prec in ("high", ""):
     global_float_prec = "double"
 elif dp_float_prec == "low":
     GLOBAL_NP_FLOAT_PRECISION = np.float32
-    GLOBAL_ENER_FLOAT_PRECISION = np.float64
+    GLOBAL_ENER_FLOAT_PRECISION = np.float32
     global_float_prec = "float"
 else:
     raise RuntimeError(
